+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Properly pick free port for error server.
 * Set a title for the corridor selector window.
 * Always pack generated resources even if packing is disabled in options.
+* Parse game translation files in an even more lenient way, to handle some Portal 2 files correctly.
 
 ------------------------------------------